requirement.txt
fastapi
uvicorn
openai
numpy
pandas
matplotlib
python-multipart
beautifulsoup4
requests
scipy
re
python-dotenv
pydantic
lxml
scikit-sklearn
parquet
DuckDB
pyarrow
openpyxl
pdfplumber
fitz
PyMuPDF
langchain
pinecone-client
io
camelot-py  # pip install camelot-py[cv]    
tabula
pytesseract
pillow
PyPDF2
seaborn
s3fs
fastparquet